Move UI files into separate directory
