+++ build_gui.sh
@@ -1,5 +1,5 @@
-/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic src/about.ui -o src/ui_about.h
-/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic src/delete_confirm.ui -o src/ui_delete_confirm.h
-/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic src/mainwindow.ui -o src/ui_mainwindow.h
-/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic src/standard_drink_calc.ui -o src/ui_standard_drink_calc.h
-/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic src/user_settings.ui -o src/ui_user_settings.h
+/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic ui/about.ui -o ui/ui_about.h
+/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic ui/delete_confirm.ui -o ui/ui_delete_confirm.h
+/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic ui/mainwindow.ui -o ui/ui_mainwindow.h
+/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic ui/standard_drink_calc.ui -o ui/ui_standard_drink_calc.h
+/Users/rwardrup/Qt/6.0.0/clang_64/bin/uic ui/user_settings.ui -o ui/ui_user_settings.h
