+++ src/about.h
@@ -5,7 +5,7 @@
 #ifndef BEERTABS_ABOUT_H
 #define BEERTABS_ABOUT_H
 
-#include "ui_about.h"
+#include "ui/ui_about.h"
 
 
 class About : public QDialog, public Ui::aboutDialog {
